Move window to another class
