+++ semester3/hw3/3.2/src/main/java/com/rprtr258/client/Entity.java
@@ -17,10 +17,6 @@ public abstract class Entity implements Renderable {
         position = pos;
     }
 
-    public Entity(double x, double y) {
-        this(new Point2D(x, y));
-    }
-
     /**
      * Adds velocity allowing to move entity.
      * @param delta vector to add
